Move a bunch of stuff around and split up big files
